restruct
move cehck_download_profess from download.py to utils.py | move all metadata func to new directory metada/set.py
